Turn on nvidia mode only in desktop mode
Clickable detects an nvidia environment automatically and turns on nvidia mode in that case.
It should have done that only when `clickable desktop` was run, but it also did this for all other modes, which is wrong.
There are no `*-nvidia` images for these modes (i.e. no `armhf-nvidia` image, which wouldn't make sense anyways).